openvpn: remove legacy resin-api routes and configuration
The `duplicate-cn` option and the "10.255.255.0" subnet were used when
the the resin-api instances were directly connecting to the VPN in order
to talk to devices.

Port 1194 was used by the legacy VPN server which is now removed. No
need to need to expose it from the container.

The iptables rules were only there to prevent client-to-client
communication, which is only enabled by the `client-to-client` directive
in the openvpn configuration. This option is no longer present.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>