openvpn: set the gateway ip to the value supplied by etcd
The gateway ip is what ends up in the routing table of every connecting
client. This allows the client to communicate with the openvpn server
over the tun interface.

By setting this IP to be a public IP allocated by us we prevent causing
any routing table collisions that could happen if the client and the VPN
server were sharing the same subnet and gateway.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>