Remove legacy activePort data
activePort was needed when we allowed multiple connections with the same
uuid, which isn't needed anymore as discussed on flowdock.

If the service is restarted it causes devices to not be set as
disconnected as the active ports are lost. The same bug could also be triggered in case of a device
disconnecting and reconnecting immediately (we have thousands of
connection-reset logs per day that could trigger it).