service: Ensure captureException does not throw
The code in the captureException() method 'try's
to get an ID. If it fails, it 'finally' registers
the exception passed to it. Unfortunately, because
it doesn't 'catch', if the 'try' fails, then the
error from it bubbles up and the code calling
'captureException' does not finalise. This is
most prevalent should the API not currently be
running, as it won't register and the timed retry
to register will never get called.

This will also fix other calls made with it.

Change-type: patch
Signed-off-by: Heds Simons <heds@balena.io>